04/03/2022

+++ tutoForum/AAE-my-questions.php
@@ -25,7 +25,7 @@ include '../tutoForum/includes/head.php';
                     <div class="title-subject">
                         <h5>Titre : <?= $question['titre']; ?></h5>
                     </div>
-                    <div class="content-subject">
+                    <div class="description-subject">
                         <p>Description : <?= $question['description'] ?></p>
                     </div>
                     <div class="authorDate-subject">
@@ -34,7 +34,7 @@ include '../tutoForum/includes/head.php';
                     <div class="modif-subject row centerAlign centerJust">
                         <a class="btn-afficher btn-green neumorph-btn row centerJust centerAlign" href="../tutoForum/CD-f-show-topics.php?titre=<?= url_custom_encode($question['titre']) ?>&id=<?= $question['id']; ?>">AFFICHER</a>
                         <a class="btn-modifier btn-green neumorph-btn row centerJust centerAlign" href="../tutoForum/AAF-edit-question.php?id=<?= $question['id'] ?>">MODIFIER</a>
-                        <a class="btn-effacer row centerJust centerAlign" href="../tutoForum/actions/topic/deleteQuestionAction.php?id=<?= $question['id'] ?>">EFFACER</a>
+                        <a class="btn-effacer row centerJust centerAlign" onclick='return confirm(" Cette action est irréversible, êtes-vous sûr ? ")' href="../tutoForum/actions/topic/deleteQuestionAction.php?id=<?= $question['id'] ?>">EFFACER</a>
                     </div>
                 </div>
             </div>
